backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2 

# Elasticsearch integration
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0

# Utilities
python-magic>=0.4.27  # for file type detection
django-cleanup>=8.0.0  # for automatic file cleanup
django-filter>=23.0  # for filtering in DRF

# Performance and monitoring
django-debug-toolbar>=4.0.0  # for development debugging 